Lift state from Board to Game
